OPENMEETINGS-722 minor changes.
git-svn-id: https://svn.apache.org/repos/asf/openmeetings/trunk/singlewebapp@1509073 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/openmeetings/web/pages/auth/RegisterDialog.java
@@ -96,7 +96,7 @@ public class RegisterDialog extends AbstractFormDialog<String> {
 		this.s = s;
 	}
 
-	public void setBrowserTZOffset(AjaxRequestTarget target, double browserTZOffset) {
+	public void setBrowserTZOffset(AjaxRequestTarget target, int browserTZOffset) {
 		tzModel.setObject(WebSession.get().getOmTimeZoneByBrowserLocale(browserTZOffset));
 		target.add(tzDropDown);
 	}
